requirements.txt
mysql-connector-python>=8.0.0
sqlalchemy>=2.0.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
